chore: removing component from entity
Entity classes are supposed to be session-bounded. By annotating a class with @Component you're flagging it for spring to instantiate it as a managed bean, keeping it as a singleton.

As an entity is supposed to be just a data holder for the database, it makes no sense to use it as a component.

The scope controls the component scope or in simple words, marks how spring will control this singleton lifecycle. You can set a scope per user session (meaning that a bean will exist for each user session) or more global.

Fix #188