feat(FSADT1-1363): backend fuzzy match for individuals (#996)
* feat(FSADT1-1363): adding new endpoint and cleaning old code

* chore(FSADT1-1363): changing testcontainer for oracle

changing it due to the issues with Mac M2 chip. Increased the timeout to allow the oracle instance to spin up properly

* chore(FSADT1-1363): updating test data

updating test data for oracle

* feat(FSADT1-1363): fuzzy match for individuals

initial commit with some of the heavy lifting. not final

* chore(FSADT1-1363): adding temporary api auth

this entry is temporary and it will be changed before final push

* feat(FSADT1-1363): updated api auth to matches

* chore(FSADT1-1363): refactoring matcher code

Refactored the code to allow easy to use on next iterations

* test(FSADT1-1363): adding tests

* chore: code cleanup

* feat(FSADT1-1363): setting flag to allow match endpoint to work

* feat(FSADT1-1363): add fe stub

* chore(FSADT1-1363): code refactor

* chore(FSADT1-1363): updating oracle sql files

* fix(FSADT1-1363): mergind and rebasing

* fix(FSADT1-1363): fixed branch due to dto changes

* test(FSADT1-1363): updating test