feat(FSADT1-1388): added match notification
This can serve as a base for the fuzzy match notification.

The FuzzyMatchNotificationGroupingComponent.vue is the responsible for showing the cds-actionable-notification. It has an ID and it should react/show data only if the received notification is marked for it. This way we can have as many as required throught the entire page and only the one that matches the Id will pop-up.

New types were added to match the fuzzy response and some other fuzzy related data.

An example on how to use it globally can be seen on the FormStaffPage.vue file. <fuzzy-match-notification-grouping-component id="global" /> sets the "global" one that will apear on top and show a notification, such as for BC Registrered ones.

To trigger it check the usage on the BcRegisteredClientInformationWizardStep.vue file around line 172~184. The errorBus.emit is to force the field to show the error message below it. The fuzzyBus is to force the fuzzy notification to appear. The ID used here was the global, so it should make the global one appear only, no other notification should appear.