feat(FSADT1-1388): Create client for staff Step 1 - BC Registered Business (#1056)
* chore: updating config

* feat(FSADT1-1388): updated readonly component

* feat(FSADT1-1388): added bc registered staff loading

* docs: updated stubs

* chore: added validation to isNotEmpty to check if string

* feat(FSADT1-1388): initial version of bc registered

* chore: added new stub

* feat(FSADT1-1388): added bc reg validations

* feat(FSADT1-1388): added validations to bcreg

* test(FSADT1-1388): updated tests and cypress files

* chore: added IDs to error notification

* chore: small update on submission validator

* feat(FSADT1-1388): added match notification

This can serve as a base for the fuzzy match notification.

The FuzzyMatchNotificationGroupingComponent.vue is the responsible for showing the cds-actionable-notification. It has an ID and it should react/show data only if the received notification is marked for it. This way we can have as many as required throught the entire page and only the one that matches the Id will pop-up.

New types were added to match the fuzzy response and some other fuzzy related data.

An example on how to use it globally can be seen on the FormStaffPage.vue file. <fuzzy-match-notification-grouping-component id="global" /> sets the "global" one that will apear on top and show a notification, such as for BC Registrered ones.

To trigger it check the usage on the BcRegisteredClientInformationWizardStep.vue file around line 172~184. The errorBus.emit is to force the field to show the error message below it. The fuzzyBus is to force the fuzzy notification to appear. The ID used here was the global, so it should make the global one appear only, no other notification should appear.

* feat(FSADT1-1388): added bc registered handling for all types

added the ability to handle duplicated and enabled the screen to handle possible external invalid cases.

* test: updated cypress fixtures and commands

* chore: updated staff validation config

* test(FSADT1-1388): added more tests

* feat(FSADT1-1388): allowing idir to receive responses

* test(FSADT1-1388): added test case for 404

* feat(FSADT1-1388): added work safe bc to legacy

* fix(FSADT1-1388): added workSafeBCNumber to dto

* fix(FSADT1-1388): fixed fuzzy notification

* feat(FSADT1-1388): added registered SP staff variation

* feat(FSADT1-1388): broaden the doing business as code

* chore(FSADT1-1388): added test case name to test

* chore: disabling content of another feature

* chore: returned value

* test: fixing tests

* fix: fixing tests

* test: fixing test

* test: fixing test

* test: fixing tests

* test: fixing test