fix(FSADT1-1435): updated staff to not rely on field ids
This change removes the reliance on the list of fields/checks and force the revalidation to be executed for all fields as requested by the bus.

This bus will request all fields to revalidate, but only visible fields are affected, as fields not visible/enabled by vue will not be there to receive the request from the bus.

The main difference now is just on the last step, as it will not have the complete revalidation, but for the staff page is ok, as there will be the fuzzy in between each step