fix(FSADT1-1737): fixing contact edit issue
Due to the nature of the contact data model, there was an issue that caused contacts to be duplicated in the frontend if it wasn't properly handled. The issue was related to name changes, as this is what "tie together" multiple contacts associated to multiple locations.

As the data model don't make use of a one-to-many association in the database, it ends up duplicating entries while changing a single association column, and to handle that it needs to group again the contacts based on name.

When a name was changed, it used to update just the selected contact by its id, ignoring any other "similar" contacts, breaking the association.

We fixed that with the duplication of the patch data before sending it to the actual update process.